feat: ajouter datadog logger avec json format
